+++ packages/xchain-bitcoin/src/client.ts
@@ -14,6 +14,7 @@ import {
   TxsPage,
   UTXOClient,
   XChainClientParams,
+  checkFeeBounds,
 } from '@xchainjs/xchain-client'
 import { getSeed } from '@xchainjs/xchain-crypto'
 import {
@@ -27,7 +28,7 @@ import {
 } from '@xchainjs/xchain-util'
 import * as Bitcoin from 'bitcoinjs-lib'
 
-import { BTC_DECIMAL } from './const'
+import { BTC_DECIMAL, LOWER_FEE_BOUND, UPPER_FEE_BOUND } from './const'
 import * as sochain from './sochain-api'
 import { ClientUrl } from './types/client-types'
 import * as Utils from './utils'
@@ -52,6 +53,10 @@ class Client extends UTXOClient {
    */
   constructor({
     network = Network.Testnet,
+    feeBounds = {
+      lower: LOWER_FEE_BOUND,
+      upper: UPPER_FEE_BOUND,
+    },
     sochainUrl = 'https://sochain.com/api/v2',
     haskoinUrl = {
       [Network.Testnet]: 'https://api.haskoin.com/btctest',
@@ -65,7 +70,7 @@ class Client extends UTXOClient {
     },
     phrase = '',
   }: BitcoinClientParams) {
-    super(Chain.Bitcoin, { network, rootDerivationPaths, phrase })
+    super(Chain.Bitcoin, { network, rootDerivationPaths, phrase, feeBounds })
     this.setSochainUrl(sochainUrl)
     this.haskoinUrl = haskoinUrl
   }
@@ -290,12 +295,15 @@ class Client extends UTXOClient {
    *
    * @param {TxParams&FeeRate} params The transfer options.
    * @returns {TxHash} The transaction hash.
+   *
+   * @throws {"memo too long"} Thrown if memo longer than  80 chars.
    */
   async transfer(params: TxParams & { feeRate?: FeeRate }): Promise<TxHash> {
     const fromAddressIndex = params?.walletIndex || 0
 
     // set the default fee rate to `fast`
     const feeRate = params.feeRate || (await this.getFeeRates())[FeeOption.Fast]
+    checkFeeBounds(this.feeBounds, feeRate)
 
     /**
      * do not spend pending UTXOs when adding a memo
@@ -331,6 +339,7 @@ class Client extends UTXOClient {
    *
    * @throws {"halted chain"} Thrown if chain is halted.
    * @throws {"halted trading"} Thrown if trading is halted.
+   * @throws {"memo too long"} Thrown if memo longer than  80 chars.
    */
   async deposit({ walletIndex = 0, asset = AssetBTC, amount, memo }: DepositParams): Promise<TxHash> {
     const inboundDetails = await getInboundDetails(asset.chain, this.network)