+++ packages/xchain-bitcoin/src/const.ts
@@ -6,3 +6,6 @@
 export const MIN_TX_FEE = 1000
 
 export const BTC_DECIMAL = 8
+
+export const LOWER_FEE_BOUND = 1
+export const UPPER_FEE_BOUND = 500