+++ packages/xchain-bitcoincash/src/const.ts
@@ -1,0 +1,2 @@
+export const LOWER_FEE_BOUND = 1
+export const UPPER_FEE_BOUND = 500