+++ packages/xchain-client/src/feeRates.ts
@@ -1,4 +1,4 @@
-import { FeeOption, FeeRate, FeeRates } from './types'
+import { FeeOption, FeeRate, FeeRates, FeeBounds } from './types'
 
 export function singleFeeRate(rate: FeeRate): FeeRates {
   return Object.values(FeeOption).reduce<Partial<FeeRates>>((a, x) => ((a[x] = rate), a), {}) as FeeRates
@@ -9,5 +9,11 @@ export function standardFeeRates(rate: FeeRate): FeeRates {
     ...singleFeeRate(rate),
     [FeeOption.Average]: rate * 0.5,
     [FeeOption.Fastest]: rate * 5.0,
+  }
+}
+
+export function checkFeeBounds(feeBounds: FeeBounds, feeRate: FeeRate): void {
+  if (feeRate < feeBounds.lower || feeRate > feeBounds.upper) {
+    throw Error(`Fee outside of predetermined bounds: ${feeRate.toString()}`)
   }
 }