+++ packages/xchain-ethereum/src/client.ts
@@ -16,6 +16,7 @@ import {
   TxsPage,
   XChainClient,
   XChainClientParams,
+  checkFeeBounds,
   standardFeeRates,
 } from '@xchainjs/xchain-client'
 import {
@@ -28,9 +29,10 @@ import {
   delay,
   getInboundDetails,
 } from '@xchainjs/xchain-util'
-import { BigNumber, BigNumberish, Wallet, ethers } from 'ethers'
+import { BigNumber, Wallet, ethers } from 'ethers'
 import { HDNode, parseUnits, toUtf8Bytes } from 'ethers/lib/utils'
 
+import { LOWER_FEE_BOUND, UPPER_FEE_BOUND } from './const'
 import erc20ABI from './data/erc20.json'
 import routerABI from './data/routerABI.json'
 import * as etherscanAPI from './etherscan-api'
@@ -53,14 +55,20 @@ import {
   BASE_TOKEN_GAS_COST,
   ETHAddress,
   ETH_DECIMAL,
-  MAX_APPROVAL,
   SIMPLE_GAS_COST,
+  call,
+  estimateApprove,
+  estimateCall,
+  getApprovalAmount,
+  getDecimal,
   getDefaultGasPrices,
   getFee,
   getTokenAddress,
   getTokenBalances,
   getTxFromEthplorerEthTransaction,
   getTxFromEthplorerTokenOperation,
+  isApproved,
+  strip0x,
   validateAddress,
   xchainNetworkToEths,
 } from './utils'
@@ -79,6 +87,7 @@ export interface EthereumClient {
   approve(params: ApproveParams): Promise<TransactionResponse>
   // `getFees` of `BaseXChainClient` needs to be overridden
   getFees(params: TxParams): Promise<Fees>
+  getWallet(walletIndex?: number): ethers.Wallet
 }
 
 export type EthereumClientParams = XChainClientParams & {
@@ -108,6 +117,10 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    */
   constructor({
     network = Network.Testnet,
+    feeBounds = {
+      lower: LOWER_FEE_BOUND,
+      upper: UPPER_FEE_BOUND,
+    },
     ethplorerUrl = 'https://api.ethplorer.io',
     ethplorerApiKey = 'freekey',
     explorerUrl,
@@ -120,7 +133,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
     etherscanApiKey,
     infuraCreds,
   }: EthereumClientParams) {
-    super(Chain.Ethereum, { network, rootDerivationPaths })
+    super(Chain.Ethereum, { network, rootDerivationPaths, feeBounds })
     this.ethNetwork = xchainNetworkToEths(network)
     this.infuraCreds = infuraCreds
     this.etherscanApiKey = etherscanApiKey
@@ -350,13 +363,14 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
         // use etherscan for testnet
 
         const newAssets = assets || [AssetETH]
+        const provider = this.getProvider()
         // Follow approach is only for testnet
         // For mainnet, we will use ethplorer api(one request only)
         // https://github.com/xchainjs/xchainjs-lib/issues/252
         // And to avoid etherscan api call limit, it gets balances in a sequence way, not in parallel
         const balances = []
         for (let i = 0; i < newAssets.length; i++) {
-          const asset = newAssets[i]
+          const asset: Asset = newAssets[i]
           const etherscan = this.getEtherscanProvider()
           if (assetToString(asset) !== assetToString(AssetETH)) {
             // Handle token balances
@@ -370,10 +384,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
               assetAddress,
               apiKey: etherscan.apiKey,
             })
-            const decimals =
-              BigNumber.from(
-                await this.call<BigNumberish>({ contractAddress: assetAddress, abi: erc20ABI, funcName: 'decimals' }),
-              ).toNumber() || ETH_DECIMAL
+            const decimals = (await getDecimal(asset, provider)) || ETH_DECIMAL
 
             if (!Number.isNaN(decimals)) {
               balances.push({
@@ -452,15 +463,15 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * @throws {"router address is not defined"} Thrown if router address is not defined
    */
   async deposit({ walletIndex = 0, asset = AssetETH, amount, memo }: DepositParams): Promise<TxHash> {
-    const inboundDetails = await getInboundDetails(asset.chain, this.network)
+    const { haltedChain, haltedTrading, router, vault } = await getInboundDetails(asset.chain, this.network)
 
-    if (inboundDetails.haltedChain) {
+    if (haltedChain) {
       throw new Error(`Halted chain for ${assetToString(asset)}`)
     }
-    if (inboundDetails.haltedTrading) {
+    if (haltedTrading) {
       throw new Error(`Halted trading for ${assetToString(asset)}`)
     }
-    if (!inboundDetails.router) {
+    if (!router) {
       throw new Error('router address is not defined')
     }
 
@@ -468,9 +479,9 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
     const gasPrice = await this.estimateGasPrices()
 
     if (asset.ticker.toUpperCase() === 'ETH') {
-      const contract = new ethers.Contract(inboundDetails.router, routerABI)
+      const contract = new ethers.Contract(router, routerABI)
       const unsignedTx = await contract.populateTransaction.deposit(
-        inboundDetails.vault,
+        vault,
         ETHAddress,
         amount.amount().toFixed(),
         memo,
@@ -484,28 +495,28 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
       return typeof response === 'string' ? response : response.hash
     } else {
       const assetAddress = asset.symbol.slice(asset.ticker.length + 1)
-      const strip0x = assetAddress.substr(2)
-      const isApproved = await this.isApproved({
+      const contractAddress = strip0x(assetAddress)
+      const isApprovedResult = await this.isApproved({
         amount: baseAmount(amount.amount()),
-        spenderAddress: inboundDetails.router,
-        contractAddress: strip0x,
+        spenderAddress: router,
+        contractAddress,
         walletIndex,
       })
 
-      if (!isApproved) {
+      if (!isApprovedResult) {
         throw new Error('The amount is not allowed to spend')
       }
 
-      const checkSummedAddress = ethers.utils.getAddress(strip0x)
-      const params = [inboundDetails.vault, checkSummedAddress, amount.amount().toFixed(), memo]
-      const vaultContract = new ethers.Contract(inboundDetails.router, routerABI)
+      const checkSummedContractAddress = ethers.utils.getAddress(contractAddress)
+      const params = [vault, checkSummedContractAddress, amount.amount().toFixed(), memo]
+      const vaultContract = new ethers.Contract(router, routerABI)
       const unsignedTx = await vaultContract.populateTransaction.deposit(...params, {
         from: address,
         value: 0,
         gasPrice: gasPrice.fast.amount().toFixed(),
       })
-      const response = await this.getWallet(walletIndex).sendTransaction(unsignedTx)
-      return typeof response === 'string' ? response : response.hash
+      const { hash } = await this.getWallet(walletIndex).sendTransaction(unsignedTx)
+      return hash
     }
   }
 
@@ -569,22 +580,18 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
 
   /**
    * Call a contract function.
-   * @template T The result interface.
-   * @param {number} walletIndex (optional) HD wallet index
+
+   * @param {signer} Signer of the transaction (optional - needed for sending transactions only)
    * @param {Address} contractAddress The contract address.
    * @param {ContractInterface} abi The contract ABI json.
    * @param {string} funcName The function to be called.
-   * @param {any[]} funcParams The parameters of the function.
-   * @returns {T} The result of the contract function call.
+   * @param {unknown[]} funcParams (optional) The parameters of the function.
    *
-   * @throws {"contractAddress must be provided"}
-   * Thrown if the given contract address is empty.
+   * @returns {T} The result of the contract function call.
    */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  async call<T>({ walletIndex = 0, contractAddress, abi, funcName, funcParams = [] }: CallParams): Promise<T> {
-    if (!contractAddress) throw new Error('contractAddress must be provided')
-    const contract = new ethers.Contract(contractAddress, abi, this.getProvider()).connect(this.getWallet(walletIndex))
-    return contract[funcName](...funcParams)
+  async call<T>({ signer, contractAddress, abi, funcName, funcParams = [] }: CallParams): Promise<T> {
+    const provider = this.getProvider()
+    return call({ provider, signer, contractAddress, abi, funcName, funcParams })
   }
 
   /**
@@ -594,21 +601,18 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * @param {string} funcName The function to be called.
    * @param {any[]} funcParams The parameters of the function.
    * @param {number} walletIndex (optional) HD wallet index
-   * @returns {BigNumber} The result of the contract function call.
    *
-   * @throws {"contractAddress must be provided"}
-   * Thrown if the given contract address is empty.
+   * @returns {BigNumber} The result of the contract function call.
+
    */
-  async estimateCall({
-    contractAddress,
-    abi,
-    funcName,
-    funcParams = [],
-    walletIndex = 0,
-  }: EstimateCallParams): Promise<BigNumber> {
-    if (!contractAddress) throw new Error('contractAddress must be provided')
-    const contract = new ethers.Contract(contractAddress, abi, this.getProvider()).connect(this.getWallet(walletIndex))
-    return contract.estimateGas[funcName](...funcParams)
+  async estimateCall({ contractAddress, abi, funcName, funcParams = [] }: EstimateCallParams): Promise<BigNumber> {
+    return estimateCall({
+      provider: this.getProvider(),
+      contractAddress,
+      abi,
+      funcName,
+      funcParams,
+    })
   }
 
   /**
@@ -620,24 +624,23 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * @param {number} walletIndex (optional) HD wallet index
    * @returns {boolean} `true` or `false`.
    */
-  async isApproved({ contractAddress, spenderAddress, amount, walletIndex = 0 }: IsApprovedParams): Promise<boolean> {
-    // since amount is optional, set it to smallest amount by default
-    const txAmount = BigNumber.from(amount?.amount().toFixed() ?? 1)
-    const owner = this.getAddress(walletIndex)
-    const allowance = await this.call<BigNumberish>({
+  async isApproved({ contractAddress, spenderAddress, amount, walletIndex }: IsApprovedParams): Promise<boolean> {
+    const allowance = await isApproved({
+      provider: this.getProvider(),
+      amount,
+      spenderAddress,
       contractAddress,
-      abi: erc20ABI,
-      funcName: 'allowance',
-      funcParams: [owner, spenderAddress],
+      fromAddress: this.getAddress(walletIndex),
     })
-    return txAmount.lte(allowance)
-  }
 
+    return allowance
+  }
   /**
    * Check allowance.
    *
    * @param {Address} contractAddress The contract address.
    * @param {Address} spenderAddress The spender address.
+   * @param {Address} fromAddress The address a transaction is send from.
    * @param {feeOptionKey} FeeOption Fee option (optional)
    * @param {BaseAmount} amount The amount of token. By default, it will be unlimited token allowance. (optional)
    * @param {number} walletIndex (optional) HD wallet index
@@ -661,48 +664,52 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
         .amount()
         .toFixed(),
     )
+
+    const fromAddress = this.getAddress(walletIndex)
+
     const gasLimit = await this.estimateApprove({
-      walletIndex,
       spenderAddress,
       contractAddress,
+      walletIndex,
       amount,
     }).catch(() => BigNumber.from(gasLimitFallback))
 
-    const txAmount = amount ? BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
+    const txAmount: BigNumber = getApprovalAmount(amount)
+    checkFeeBounds(this.feeBounds, gasPrice.toNumber())
+
     return await this.call<TransactionResponse>({
-      walletIndex,
+      signer: this.getWallet(walletIndex),
       contractAddress,
       abi: erc20ABI,
       funcName: 'approve',
-      funcParams: [spenderAddress, txAmount, { from: this.getAddress(walletIndex), gasPrice, gasLimit }],
+      funcParams: [spenderAddress, txAmount, { from: fromAddress, gasPrice, gasLimit }],
     })
   }
 
   /**
-   * Estimate gas limit of approve.
+   * Estimate gas for calling `approve`.
    *
    * @param {Address} contractAddress The contract address.
    * @param {Address} spenderAddress The spender address.
-   * @param {number} walletIndex (optional) HD wallet index
+   * @param {Address} fromAddress The spender address.
    * @param {BaseAmount} amount The amount of token. By default, it will be unlimited token allowance. (optional)
-   * @returns {BigNumber} The estimated gas limit.
+   *
+   * @returns {BigNumber} Estimated gas
    */
   async estimateApprove({
+    walletIndex = 0,
     contractAddress,
     spenderAddress,
-    walletIndex = 0,
     amount,
   }: EstimateApproveParams): Promise<BigNumber> {
-    const txAmount = amount ? BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
-    const gasLimit = await this.estimateCall({
-      walletIndex,
+    return await estimateApprove({
+      provider: this.getProvider(),
       contractAddress,
+      spenderAddress,
+      fromAddress: this.getAddress(walletIndex),
       abi: erc20ABI,
-      funcName: 'approve',
-      funcParams: [spenderAddress, txAmount, { from: this.getAddress(walletIndex) }],
+      amount,
     })
-
-    return gasLimit
   }
 
   /**
@@ -762,11 +769,15 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
       }
     }
 
-    let txResult
+    checkFeeBounds(this.feeBounds, BigNumber.from(overrides.gasPrice).toNumber())
+
+    const signer = this.getWallet(walletIndex)
+
+    let txResult: TransactionResponse
     if (assetAddress && !isETHAddress) {
       // Transfer ERC20
       txResult = await this.call<TransactionResponse>({
-        walletIndex,
+        signer,
         contractAddress: assetAddress,
         abi: erc20ABI,
         funcName: 'transfer',
@@ -782,7 +793,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
         },
       )
 
-      txResult = await this.getWallet().sendTransaction(transactionRequest)
+      txResult = await signer.sendTransaction(transactionRequest)
     }
 
     return txResult.hash
@@ -810,7 +821,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
     try {
       return await this.estimateGasPricesFromEtherscan()
     } catch (error) {
-      return Promise.reject(new Error(`Failed to estimate gas price: ${error.msg ?? error.toString()}`))
+      return Promise.reject(new Error(`Failed to estimate gas price: ${error}`))
     }
   }