+++ packages/xchain-ethereum/src/const.ts
@@ -1,0 +1,2 @@
+export const LOWER_FEE_BOUND = 2_000_000_000
+export const UPPER_FEE_BOUND = 1_000_000_000_000