+++ packages/xchain-ethereum/src/utils.ts
@@ -10,7 +10,7 @@ import {
   assetToString,
   baseAmount,
 } from '@xchainjs/xchain-util'
-import { BigNumber, ethers, providers } from 'ethers'
+import { Signer, ethers, providers } from 'ethers'
 import { parseUnits } from 'ethers/lib/utils'
 
 import erc20ABI from './data/erc20.json'
@@ -31,14 +31,14 @@ export const ETHPLORER_FREEKEY = 'freekey'
 
 // from https://github.com/MetaMask/metamask-extension/blob/ee205b893fe61dc4736efc576e0663189a9d23da/ui/app/pages/send/send.constants.js#L39
 // and based on recommendations of https://ethgasstation.info/blog/gas-limit/
-export const SIMPLE_GAS_COST: ethers.BigNumber = BigNumber.from(21000)
-export const BASE_TOKEN_GAS_COST: ethers.BigNumber = BigNumber.from(100000)
+export const SIMPLE_GAS_COST: ethers.BigNumber = ethers.BigNumber.from(21000)
+export const BASE_TOKEN_GAS_COST: ethers.BigNumber = ethers.BigNumber.from(100000)
 
 // default gas price in gwei
 export const DEFAULT_GAS_PRICE = 50
 
 export const ETHAddress = '0x0000000000000000000000000000000000000000'
-export const MAX_APPROVAL = BigNumber.from(2).pow(256).sub(1)
+export const MAX_APPROVAL: ethers.BigNumber = ethers.BigNumber.from(2).pow(256).sub(1)
 
 /**
  * Network -> EthNetwork
@@ -241,7 +241,7 @@ export const getTxFromEthplorerEthTransaction = (txInfo: TransactionInfo): Tx =>
  *
  * @returns {Fees} The default gas price.
  */
-export const getFee = ({ gasPrice, gasLimit }: { gasPrice: BaseAmount; gasLimit: BigNumber }) =>
+export const getFee = ({ gasPrice, gasLimit }: { gasPrice: BaseAmount; gasLimit: ethers.BigNumber }) =>
   baseAmount(gasPrice.amount().multipliedBy(gasLimit.toString()), ETH_DECIMAL)
 
 export const estimateDefaultFeesWithGasPricesAndLimits = (asset?: Asset): FeesWithGasPricesAndLimits => {
@@ -259,9 +259,9 @@ export const estimateDefaultFeesWithGasPricesAndLimits = (asset?: Asset): FeesWi
 
   let gasLimit
   if (assetAddress && assetAddress !== ETHAddress) {
-    gasLimit = BigNumber.from(BASE_TOKEN_GAS_COST)
+    gasLimit = ethers.BigNumber.from(BASE_TOKEN_GAS_COST)
   } else {
-    gasLimit = BigNumber.from(SIMPLE_GAS_COST)
+    gasLimit = ethers.BigNumber.from(SIMPLE_GAS_COST)
   }
 
   return {
@@ -323,6 +323,113 @@ export const filterSelfTxs = <T extends { from: string; to: string; hash: string
 }
 
 /**
+ * Returns approval amount
+ *
+ * If given amount is not set or zero, `MAX_APPROVAL` amount is used
+ */
+export const getApprovalAmount = (amount?: BaseAmount): ethers.BigNumber =>
+  amount && amount.gt(baseAmount(0, amount.decimal)) ? ethers.BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
+
+/**
+ * Call a contract function.
+ *
+ * @param {Provider} provider Provider to interact with the contract.
+ * @param {Address} contractAddress The contract address.
+ * @param {ContractInterface} abi The contract ABI json.
+ * @param {string} funcName The function to be called.
+ * @param {unknown[]} funcParams The parameters of the function.
+ * @returns {BigNumber} The result of the contract function call.
+ */
+export const estimateCall = async ({
+  provider,
+  contractAddress,
+  abi,
+  funcName,
+  funcParams = [],
+}: {
+  provider: providers.Provider
+  contractAddress: Address
+  abi: ethers.ContractInterface
+  funcName: string
+  funcParams?: unknown[]
+}): Promise<ethers.BigNumber> => {
+  const contract: ethers.Contract = new ethers.Contract(contractAddress, abi, provider)
+  return await contract.estimateGas[funcName](...funcParams)
+}
+
+/**
+ * Calls a contract function.
+ *
+ * @param {Provider} provider Provider to interact with the contract.
+ * @param {signer} Signer of the transaction (optional - needed for sending transactions only)
+ * @param {Address} contractAddress The contract address.
+ * @param {ContractInterface} abi The contract ABI json.
+ * @param {string} funcName The function to be called.
+ * @param {unknow[]} funcParams (optional) The parameters of the function.
+ *
+ * @returns {T} The result of the contract function call.
+
+ */
+export const call = async <T>({
+  provider,
+  signer,
+  contractAddress,
+  abi,
+  funcName,
+  funcParams = [],
+}: {
+  provider: providers.Provider
+  signer?: Signer
+  contractAddress: Address
+  abi: ethers.ContractInterface
+  funcName: string
+  funcParams?: unknown[]
+}): Promise<T> => {
+  let contract = new ethers.Contract(contractAddress, abi, provider)
+  if (signer) {
+    // For sending transactions a signer is needed
+    contract = contract.connect(signer)
+  }
+  return contract[funcName](...funcParams)
+}
+
+/**
+ * Estimate gas for calling `approve`.
+ *
+ * @param {Provider} provider Provider to interact with the contract.
+ * @param {Address} contractAddress The contract address.
+ * @param {Address} spenderAddress The spender address.
+ * @param {Address} fromAddress The address a transaction is sent from.
+ * @param {BaseAmount} amount (optional) The amount of token. By default, it will be unlimited token allowance.
+ *
+ * @returns {BigNumber} Estimated gas
+ */
+export const estimateApprove = async ({
+  provider,
+  contractAddress,
+  spenderAddress,
+  fromAddress,
+  abi,
+  amount,
+}: {
+  provider: providers.Provider
+  contractAddress: Address
+  spenderAddress: Address
+  fromAddress: Address
+  abi: ethers.ContractInterface
+  amount?: BaseAmount
+}): Promise<ethers.BigNumber> => {
+  const txAmount = getApprovalAmount(amount)
+  return await estimateCall({
+    provider,
+    contractAddress,
+    abi,
+    funcName: 'approve',
+    funcParams: [spenderAddress, txAmount, { from: fromAddress }],
+  })
+}
+
+/**
  * Get Decimals
  *
  * @param {Asset} asset
@@ -340,6 +447,37 @@ export const getDecimal = async (asset: Asset, provider: providers.Provider): Pr
   const decimal: ethers.BigNumberish = await contract.decimals()
 
   return ethers.BigNumber.from(decimal).toNumber()
+}
+
+/**
+ * Check allowance.
+ *
+ * @param {Provider} provider Provider to interact with the contract.
+ * @param {Address} contractAddress The contract (ERC20 token) address.
+ * @param {Address} spenderAddress The spender address (router).
+ * @param {Address} fromAddress The address a transaction is sent from.
+ * @param {BaseAmount} amount The amount to check if it's allowed to spend or not (optional).
+ * @param {number} walletIndex (optional) HD wallet index
+ * @returns {boolean} `true` or `false`.
+ */
+export const isApproved = async ({
+  provider,
+  contractAddress,
+  spenderAddress,
+  fromAddress,
+  amount,
+}: {
+  provider: providers.Provider
+  contractAddress: Address
+  spenderAddress: Address
+  fromAddress: Address
+  amount?: BaseAmount
+}): Promise<boolean> => {
+  const txAmount = ethers.BigNumber.from(amount?.amount().toFixed() ?? 1)
+  const contract: ethers.Contract = new ethers.Contract(contractAddress, erc20ABI, provider)
+  const allowance: ethers.BigNumberish = await contract.allowance(fromAddress, spenderAddress)
+
+  return txAmount.lte(allowance)
 }
 
 /**
@@ -369,3 +507,8 @@ export const getTokenBalances = (tokenBalances: TokenBalance[]): Balance[] => {
     return acc
   }, [] as Balance[])
 }
+
+/**
+ * Removes `0x` or `0X` from address
+ */
+export const strip0x = (addr: Address) => addr.replace(/^0(x|X)/, '')