+++ packages/xchain-litecoin/src/client.ts
@@ -14,11 +14,13 @@ import {
   TxsPage,
   UTXOClient,
   XChainClientParams,
+  checkFeeBounds,
 } from '@xchainjs/xchain-client'
 import { getSeed } from '@xchainjs/xchain-crypto'
 import { AssetLTC, Chain, assetAmount, assetToBase, assetToString, getInboundDetails } from '@xchainjs/xchain-util'
 import * as Litecoin from 'bitcoinjs-lib'
 
+import { LOWER_FEE_BOUND, UPPER_FEE_BOUND } from './const'
 import * as sochain from './sochain-api'
 import { NodeAuth } from './types'
 import { TxIO } from './types/sochain-api-types'
@@ -47,6 +49,10 @@ class Client extends UTXOClient {
    */
   constructor({
     network = Network.Testnet,
+    feeBounds = {
+      lower: LOWER_FEE_BOUND,
+      upper: UPPER_FEE_BOUND,
+    },
     sochainUrl = 'https://sochain.com/api/v2',
     phrase,
     nodeUrl,
@@ -60,7 +66,7 @@ class Client extends UTXOClient {
       [Network.Stagenet]: `m/84'/2'/0'/0/`,
     },
   }: LitecoinClientParams) {
-    super(Chain.Litecoin, { network, rootDerivationPaths, phrase })
+    super(Chain.Litecoin, { network, rootDerivationPaths, phrase, feeBounds })
     this.nodeUrl =
       nodeUrl ??
       (() => {
@@ -306,6 +312,8 @@ class Client extends UTXOClient {
   async transfer(params: TxParams & { feeRate?: FeeRate }): Promise<TxHash> {
     const fromAddressIndex = params?.walletIndex || 0
     const feeRate = params.feeRate || (await this.getFeeRates())[FeeOption.Fast]
+    checkFeeBounds(this.feeBounds, feeRate)
+
     const { psbt } = await Utils.buildTx({
       ...params,
       feeRate,