+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v.0.18.4 (2022-05-17)
+
+## Update
+
+- throw error if memo is longer than 80 chars
+
+# v.0.18.3 (2022-05-08)
+
+## Fix
+- Add improvement to error message on utxo pending
+
+
 # v.0.18.2 (2022-05-05)
 
 ## Update